COMMERCE-686 Refactor Application
